+++ demisto_sdk/__main__.py
@@ -939,25 +939,26 @@ def upload(**kwargs):
     from demisto_sdk.commands.upload.uploader import ConfigFileParser, Uploader
     from demisto_sdk.commands.zip_packs.packs_zipper import (EX_FAIL,
                                                              PacksZipper)
-    if kwargs['zip'] or kwargs['input_config_file']:
-        if kwargs.pop('zip', False):
+    keep_zip = kwargs.pop('keep_zip')
+    is_zip = kwargs.pop('zip', False)
+    config_file_path = kwargs.pop('input_config_file')
+    is_xsiam = kwargs.pop('xsiam', False)
+    if is_zip or config_file_path:
+        if is_zip:
             pack_path = kwargs['input']
-            kwargs.pop('input_config_file')
 
         else:
-            config_file_path = kwargs['input_config_file']
             config_file_to_parse = ConfigFileParser(config_file_path=config_file_path)
             pack_path = config_file_to_parse.parse_file()
             kwargs['detached_files'] = True
-            kwargs.pop('input_config_file')
-        if kwargs.pop('xsiam', False):
+        if is_xsiam:
             marketplace = MarketplaceVersions.MarketplaceV2.value
         else:
             marketplace = MarketplaceVersions.XSOAR.value
         os.environ[ENV_DEMISTO_SDK_MARKETPLACE] = marketplace.lower()
 
-        output_zip_path = kwargs.pop('keep_zip') or tempfile.gettempdir()
-        packs_unifier = PacksZipper(pack_paths=pack_path, output=output_zip_path,
+        output_zip_path = keep_zip or tempfile.TemporaryDirectory()
+        packs_unifier = PacksZipper(pack_paths=pack_path, output=str(output_zip_path),
                                     content_version='0.0.0', zip_all=True, quiet_mode=True, marketplace=marketplace)
         packs_zip_path, pack_names = packs_unifier.zip_packs()
         if packs_zip_path is None and not kwargs.get('detached_files'):
@@ -965,14 +966,12 @@ def upload(**kwargs):
 
         kwargs['input'] = packs_zip_path
         kwargs['pack_names'] = pack_names
-    else:
-        kwargs.pop('zip')
-        kwargs.pop('keep_zip')
-        kwargs.pop('input_config_file')
-        kwargs.pop('xsiam', None)
 
     check_configuration_file('upload', kwargs)
-    return Uploader(**kwargs).upload()
+    upload_result = Uploader(**kwargs).upload()
+    if (is_zip or config_file_path) and not keep_zip:
+        output_zip_path.cleanup()
+    return upload_result
 
 
 # ====================== download ====================== #